meeting_logger/requirements.txt
faster-whisper
torch
openai
notion-client
python-dotenv
rich
tqdm
